pyproject.toml
[project]
name = "documentation-helper"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "black>=25.9.0",
    "certifi>=2025.10.5",
    "fastapi>=0.120.1",
    "firecrawl-py>=4.5.0",
    "ipykernel>=7.1.0",
    "isort>=7.0.0",
    "jinja2>=3.1.6",
    "langchain-anthropic>=0.3.22",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.31",
    "langgraph>=0.2.60",
    "langchain-text-splitters>=0.3.5",
    "langchain-openai>=0.2.0",
    "langchain-huggingface>=0.3.1",
    "langchain-pinecone>=0.2.12",
    "langchain-tavily>=0.2.12",
    "langchainhub>=0.1.21",
    "langsmith>=0.4.38",
    "nltk>=3.9.2",
    "openai>=2.6.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.50.0",
    "streamlit-chat>=0.1.1",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
    "unstructured>=0.18.15",
    "uvicorn>=0.38.0",
]
